Various Polish
Integrated in the opening animation which automatically loads into the first level upon hitting Start on the home screen.
Recreated the "Are you sure"/warning screen on level 10 (was missing for some reason),
Fixed the positioning of the esc button text and level number text in each level.
Fixed the scaling of the backgrounds in every level.
Renamed all level scenes to the number instead of the word, i.e. "two" changed to "2", so that they're displayed in numerical order in the project file.
Changed code to suit above.
Misc polish elements and tidying up of scenes and files.